+++ app/blueprints/registry.py
@@ -13,6 +13,7 @@ from blueprints.api.debug import debug_bp
 from blueprints.api.me import me_bp
 from blueprints.api.status import status_bp
 from blueprints.api.subscriptions import subscriptions_bp
+from blueprints.api.swagger import swagger_bp
 from blueprints.api.workflows import workflows_bp
 from blueprints.home import create_main_blueprint
 from blueprints.settings import settings_api_bp, settings_pages_bp
@@ -30,6 +31,7 @@ def register_blueprints(app: Flask, oidc: OpenIDConnect) -> None:
     register_api_blueprint(workflows_bp)
     register_api_blueprint(call_queue_bp)
     register_api_blueprint(call_session_bp)
+    register_api_blueprint(swagger_bp)
     register_api_blueprint(settings_api_bp)
     register_api_blueprint(address_search_api_bp)